closes #2 * created classes necessary for adding music folders * created rest calls for rebuilding the library * created stub html file for viewing library

+++ src/main/java/em/repos/SongInfoRepository.java
@@ -4,5 +4,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import em.model.SongInfo;
 
-public interface SongInfoRepository extends CrudRepository<SongInfo, Long> {
+/**
+ * @since v0.1
+ * @author eviljoe
+ */
+public interface SongInfoRepository extends CrudRepository<SongInfo, Integer> {
 }
